moved docmaps package up
